Move JenkinsUtil to internal package
