+++ src/main/java/eu/xenit/gradle/docker/internal/JenkinsUtil.java
@@ -1,4 +1,4 @@
-package eu.xenit.gradle;
+package eu.xenit.gradle.docker.internal;
 
 import java.text.DateFormat;
 import java.text.SimpleDateFormat;
